requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncio==3.4.3
asyncpg==0.28.0
build==1.0.3
bcrypt==4.1.2
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.70
cryptography==41.0.5
defusedxml==0.7.1
django-email-users==1.1.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exchangelib==5.1.0
fastapi==0.103.2
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
ldap3==2.9.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
oauthlib==3.2.2
orjson==3.9.7
packaging==23.2
passlib==1.7.4
pip-tools==7.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.14.5
pydentic==0.0.1.dev3
Pygments==2.17.2
pykeepass==4.0.6
pyproject_hooks==1.0.0
pysmb==1.2.9.1
PySmbClient==0.1.5
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-ldap==3.4.3
python-multipart==0.0.6
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
smbprotocol==1.12.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Ext==0.2
sqlalchemy-orm==1.2.10
starlette==0.27.0
tqdm==4.66.1
transliterate==1.10.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3

pytest==7.4.4